requirement.txt
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
youtube-transcript-api
